+++ src/assets/menus/adminMenu.tsx
@@ -1,7 +1,7 @@
 const adminNavItems = [
   {
     id: 1,
-    url: "/admin-dashboard",
+    url: "/admin/dashboard",
     label: "Dashboard",
     value: "Dashboard",
     icon: (
@@ -31,7 +31,7 @@ const adminNavItems = [
   },
   {
     id: 2,
-    url: "/admin-student-registration",
+    url: "/admin/student-registration",
     label: "Student Registration",
     value: "student-registration",
     icon: (
@@ -62,7 +62,7 @@ const adminNavItems = [
   },
   {
     id: 3,
-    url: "/admin-add-lecturer",
+    url: "/admin/add-lecturers",
     label: "Add lecturer",
     value: "add-lecturer",
     icon: (
@@ -88,7 +88,33 @@ const adminNavItems = [
   },
   {
     id: 4,
-    url: "/admin-hardware-settings",
+    url: "/admin/add-admins",
+    label: "Add Admins",
+    value: "add-admins",
+    icon: (
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        viewBox="0 0 24 24"
+        fill="currentColor"
+        className="w-6 h-6"
+      >
+        <path d="M5.25 6.375a4.125 4.125 0 1 1 8.25 0 4.125 4.125 0 0 1-8.25 0ZM2.25 19.125a7.125 7.125 0 0 1 14.25 0v.003l-.001.119a.75.75 0 0 1-.363.63 13.067 13.067 0 0 1-6.761 1.873c-2.472 0-4.786-.684-6.76-1.873a.75.75 0 0 1-.364-.63l-.001-.122ZM18.75 7.5a.75.75 0 0 0-1.5 0v2.25H15a.75.75 0 0 0 0 1.5h2.25v2.25a.75.75 0 0 0 1.5 0v-2.25H21a.75.75 0 0 0 0-1.5h-2.25V7.5Z" />
+      </svg>
+    ),
+    hasSubMenu: false,
+    subMenu: [
+      {
+        url: "#",
+        label: "",
+        value: "",
+        icon: "",
+        id: 1.1,
+      },
+    ],
+  },
+  {
+    id: 5,
+    url: "/admin/hardware-settings",
     label: "Hardware settings",
     value: "hardware-settings",
     icon: (