+++ src/pages/(onboarding)/login/lecturerLogin.tsx
@@ -10,7 +10,7 @@ import useAxios from "../../../services/base/axios/useAxios";
 import { updateAuthentication } from "../../../store/user/auth";
 import { Button, Divider } from "@mui/material";
 
-export default function Login() {
+export default function LecturerLogin() {
   const axios = useAxios();
   const dispatch = useAppDispatch();
   const navigate = useNavigate();
@@ -25,7 +25,7 @@ export default function Login() {
       e.preventDefault();
       setIsSubmitting(true);
       try {
-        const response = await axios.post("/students/login", {
+        const response = await axios.post("/lecturers/login", {
           email,
           password,
         });
@@ -39,7 +39,7 @@ export default function Login() {
           })
         );
         dispatch(openSnackbar({ message: message, isError: false }));
-        navigate("/student/dashboard");
+        navigate("/lecturer/dashboard");
       } catch (error: any) {
         const errorMessage = error?.response?.data?.message;
         dispatch(openSnackbar({ message: errorMessage, isError: true }));
@@ -63,7 +63,7 @@ export default function Login() {
               <img src="/logo192.png" alt="" className=" w-20 md:w-40 h-auto" />
             </Link>
             <h5 className=" text-2xl text-center">
-              Student User, Welcome Back
+              Lecturer User, Welcome Back
             </h5>
             <div className="w-full flex flex-col items-center justify-center gap-3">
               <form
@@ -108,10 +108,10 @@ export default function Login() {
                 <Button
                   fullWidth
                   variant="outlined"
-                  to="/lecturer/login"
+                  to="/login"
                   component={Link}
                 >
-                  Lecturer User
+                  Student User
                 </Button>
                 <Button
                   fullWidth
@@ -125,7 +125,7 @@ export default function Login() {
             </div>{" "}
           </div>
         </div>
-        <div className="w-1/2 fixed right-0 top-0  h-screen hidden md:block student-bg bg-primary-500/70 bg-center bg-no-repeat bg-cover"></div>
+        <div className="w-1/2 fixed right-0 top-0  h-screen hidden md:block lecturer-bg bg-primary-500/70 bg-center bg-no-repeat bg-cover"></div>
       </div>
     </motion.div>
   );